refactor(http): extract the team lifecycle routes from the teams router
src/main/http/teams.ts is the router every team-facing HTTP call goes
through, and it is capped at the default 800 lines with no legacy
exception to grow into. The lifecycle area is about to gain a route, and
after that there are more stop-path changes queued behind it; adding them
inline spends the router's remaining headroom on a topic that is already
self-contained.

The three routes that talk to the runtime control API - stop a team, read
one team's runtime state, list the runtime state of every alive team -
move verbatim into src/main/http/teams/teamLifecycleRoutes.ts. The bodies
are unchanged; the module takes the router's error mapping through a deps
object, exactly like registerTeamMemberDiagnosticsRoute already does, so
a lifecycle route keeps answering with the same status codes and the same
"Internal server error" redaction as its neighbours.

Sharing the mapping rather than re-deriving it is the point of the deps
object: getStatusCode encodes which errors are 400/404/409/422/501 for
this router, and a second copy of that decision is how the two halves
would drift. getTeamRuntimeApi moves with the routes because they were
its only callers.

src/main/http/teams.ts: 705 -> 642 lines.
